Corrects a clang-tidy warning about an implicit narrowing conversion from 'std::size_t' to 'int':
/home/runner/work/avogadrolibs/avogadrolibs/avogadrolibs/avogadro/core/atomutilities.cpp:70:24: warning: narrowing conversion from 'std::vector::size_type' (aka 'unsigned long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]
  int currentValence = allVectors.size();
                       ^

+++ avogadro/core/atomutilities.cpp
@@ -67,7 +67,7 @@ Vector3 AtomUtilities::generateNewBondVector(
 {
   Vector3 newPos;
   bool success = false;
-  int currentValence = allVectors.size();
+  int currentValence = static_cast<int>(allVectors.size());
 
   // No bonded atoms, just pick a random vector
   if (currentValence == 0) {
